chore(clients): move fetch handler to handlers folder
